requirements.txt
web3
pqcrypto
cffi
IPFS-Toolkit
scikit-learn
py-solc-x
torch 
torchvision
torchaudio
tenseal